requirements.txt
requests ==2.31.0
urllib3>=2.0.5,<3
aiohttp ==3.9.3
future
pydantic >=2.0.3, <3
aenum >= 3.1.11
python_dateutil >= 2.5.3
Deprecated
gunicorn
flask
line-bot-sdk
pytz
apscheduler
pypinyin